+++ Changelog.md
@@ -6,6 +6,31 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [2.5.2] - 2024-09-08
+
+### Changed
+
+- [MainFrame] Switched up the order of the info frames in the top right corner.
+
+- [Queue Dropdown] Changed the dropdown style to something similar to Blizzards new dropdowns.
+
+- [Queue Dropdown] When your character is at the last expansion's max level (e.g. 70 for dragonflight) the queue dropdown dungeon selection will be ordered by expansionlevel and Random Dungeon will always be on top.
+
+- [Queue Dropdown] When there are no options for a category type (e.g. no raid finder options) the button will be greyed out and do nothing.
+
+### Fixed
+
+- [MainFrame] Calendar events are back again with a slightly updated look.
+
+- [ActiveQueues] Hovering over most queue frames should cause a quest ID error anymore.
+
+- Clicking on the dungeon icon in the RaiderIO info panel should now refer to the correct dungeon + difficulty.
+
+- [Queue Dropdown] If you have no heroic dungeons unlocked there won't be a section in the "Specific Dungeons" selection that tells you that you have unlocked them.
+
+- [Queue Dropdown] The button to queue for multiple dungeons should now match the width of the frame.
+
+
 ## [2.5.1] - 2024-09-06
 
 ### Added
@@ -2268,7 +2293,8 @@ Basically just the rating for the bracket and the tier they're currently in
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/NintendoLink07/MythicIOGrabber/compare/2.5.1..HEAD
+[unreleased]: https://github.com/NintendoLink07/MythicIOGrabber/compare/2.5.2..HEAD
+[2.5.2]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.5.2
 [2.5.1]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.5.1
 [2.5.0]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.5.0
 [2.4.9]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.4.9